backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<11.0
pytesseract>=0.3.10,<0.4
pytz>=2023.3
dj-database-url>=3.0.1